create 'Content' module, move appropriate functions
