+++ website-engine.cabal
@@ -62,6 +62,7 @@ library
     , warnings
 
   exposed-modules:
+    Blog.Content
     Blog.Engine
     Blog.Item
     Blog.Path.Rel
@@ -82,6 +83,7 @@ library
     , lens
     , lens-aeson
     , mtl
+    , mustache
     , pandoc
     , pandoc-types
     , path
